+++ lib/camera.dart
@@ -6,7 +6,6 @@ import 'Scene.dart';
 /// A camera’s projection. A [Node] MAY reference a camera to apply a transform
 /// to place the camera in the [Scene].
 abstract class Camera extends GLTFBase {
-
   /// The user-defined name of this object. This is not necessarily unique,
   /// e.g., an accessor and a buffer could have the same name, or two accessors
   /// could even have the same name.
@@ -18,7 +17,6 @@ abstract class Camera extends GLTFBase {
 /// An orthographic camera containing properties to create an orthographic
 /// projection matrix.
 class OrthographicCamera extends Camera {
-
   /// The floating-point horizontal magnification of the view. This value
   /// MUST NOT be equal to zero. This value SHOULD NOT be negative.
   double xMag;
@@ -48,7 +46,6 @@ class OrthographicCamera extends Camera {
 /// A perspective camera containing properties to create a perspective
 /// projection matrix.
 class PerspectiveCamera extends Camera {
-
   /// The floating-point aspect ratio of the field of view. When undefined, the
   /// aspect ratio of the rendering viewport MUST be used.
   double? aspectRatio;