+++ lib/ext_texture_webp.dart
@@ -1,0 +1,7 @@
+/// https://github.com/KhronosGroup/glTF/blob/main/extensions/2.0/Vendor/EXT_texture_webp
+class EXTTextureWebp {
+  /// index of webp image
+  int source;
+
+  EXTTextureWebp({required this.source});
+}