+++ lib/khr_animation_pointer.dart
@@ -1,0 +1,5 @@
+class KHRAnimationPointer {
+  String pointer;
+
+  KHRAnimationPointer({required this.pointer});
+}