+++ lib/khr_materials_ior.dart
@@ -1,0 +1,6 @@
+class KHRMaterialIor {
+  /// The index of refraction.
+  double ior;
+
+  KHRMaterialIor({this.ior = 1.5});
+}