+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.0.3
+
+* 2.0 glTF spec conformant.
+* on demand loading assets (buffers and images)
+* implemented extensions:
+*   KHR_animation_pointer, 
+*   KHR_lights_punctual, 
+*   KHR_materials_specular, 
+*   KHR_materials_ior, 
+*   KHR_texture_transform, 
+*   EXT_texture_webp
+
 ## 0.0.2
 
 * dart format covention